Create A RestaurantNearbyBank2 where getting the Restaurant list.
Create MyJobService where using RestaurantNearbyBank2 to get the Restaurant list in background and send the list to the main thread.

Create RestaurantListManager where using MyJobService to get the Restaurant list and where getting the list from the main thread.

In RestaurantMapViewFragment getting the Restaurant list from using RestaurantListManager,
register the RestaurantListManager's broadcast in the onStart() override methode and
register the RestaurantListManager's broadcast in the onStop() override methode.